Refactoring function names. Implemented OSC communication and trial structure.
