+++ Assets/Scenes/VR_Screen/NavMesh-Screen3.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d42ff986a7d3bc04ea314b33f1114097
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
