+++ Assets/Scenes/VR_Screen/NavMesh-Screen4.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2933bc32271a3de449cec0e08274ef64
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
