+++ Assets/Scenes/VR_Screen/NavMesh-Screen5.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be53774d585689545a8a3625a2da5fc5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
