+++ Assets/Scenes/VR_preyCapture.unity
@@ -612,7 +612,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   inPort: 6969
-  outIP: 127.0.0.1
+  outIP: 127.0.0.2
   outPort: 6161
 --- !u!4 &1005790156
 Transform:
