+++ Assets/Scripts/OSC_test.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f9bc93989d8d7b4c9035474b589d3d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
